Asset price management and tradable VS non tradable assets in FTS
Added function to get symbol prices form Binance.
Asset prices added as input value to the function that initlise a local
wallet based on selected strategy.
Assets in FTS local wallet divided into those that are tradable and those
that are not. An asset is tradable if it can be converted directly to
USDT.