Analytics timestamp bug fix
The timestamp on analytics objects was set using the value found in the
local account object, which gets updated only when an operation is
issued. That caused saveral analytics objects to share the same
timestamp, which is not supposed to happen at all. The less operations
were issued, the more analytics objects would share the same timestamp.
The timestamp is now set using the value found in the price update
object.